고정 메모 섹션 수정

+++ Sesac_Memo/Extension/MemoListViewController+Extension.swift
@@ -11,14 +11,22 @@ import UIKit
 extension MemoListViewController: UITableViewDelegate, UITableViewDataSource {
     
     func numberOfSections(in tableView: UITableView) -> Int {
-        return 2
+        if tasks.filter("isFixed == true").count == 0 {
+            return 1
+        } else {
+            return 2
+        }
     }
     
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        if section == 0 {
-            return tasks.filter("isFixed == true").count
-        } else {
+        if tasks.filter("isFixed == true").count == 0 {
             return tasks.filter("isFixed == false").count
+        } else {
+            if section == 0 {
+                return tasks.filter("isFixed == true").count
+            } else {
+                return tasks.filter("isFixed == false").count
+            }
         }
     }
     
@@ -28,25 +36,31 @@ extension MemoListViewController: UITableViewDelegate, UITableViewDataSource {
         let fixedTasks = tasks.filter("isFixed == true")
         let notFixedTasks = tasks.filter("isFixed == false")
         
-        //날짜 포맷 생각하기
-        if indexPath.section == 0 {
-            cell.titleLabel.text = fixedTasks[indexPath.row].memoTitle
-            cell.bottomLabel.text = fixedTasks[indexPath.row].registerDate.checkDate() + "   " + fixedTasks[indexPath.row].memoContent
-            
-            print(indexPath)
-        } else {
+        if fixedTasks.count == 0 {
             cell.titleLabel.text = notFixedTasks[indexPath.row].memoTitle
             cell.bottomLabel.text = notFixedTasks[indexPath.row].registerDate.checkDate() + "   " + notFixedTasks[indexPath.row].memoContent
-            print(indexPath)
+        } else {
+            if indexPath.section == 0 {
+                cell.titleLabel.text = fixedTasks[indexPath.row].memoTitle
+                cell.bottomLabel.text = fixedTasks[indexPath.row].registerDate.checkDate() + "   " + fixedTasks[indexPath.row].memoContent
+            } else {
+                cell.titleLabel.text = notFixedTasks[indexPath.row].memoTitle
+                cell.bottomLabel.text = notFixedTasks[indexPath.row].registerDate.checkDate() + "   " + notFixedTasks[indexPath.row].memoContent
+            }
         }
+        
         return cell
     }
     
     func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
-        if section == 0 {
-            return "고정된 메모"
-        } else {
+        if tasks.filter("isFixed == true").count == 0 {
             return "메모"
+        } else {
+            if section == 0 {
+                return "고정된 메모"
+            } else {
+                return "메모"
+            }
         }
     }
     
@@ -61,14 +75,20 @@ extension MemoListViewController: UITableViewDelegate, UITableViewDataSource {
         let notFixedTasks = tasks.filter("isFixed == false")
         
         let delete = UIContextualAction(style: .normal, title: nil) { action, view, completionHandler in
-            if indexPath.section == 0 {
-                self.repository.deleteMemo(memo: fixedTasks[indexPath.row])
-            } else {
+            if fixedTasks.count == 0 {
                 self.repository.deleteMemo(memo: notFixedTasks[indexPath.row])
+            } else {
+                if indexPath.section == 0 {
+                    self.repository.deleteMemo(memo: fixedTasks[indexPath.row])
+                } else {
+                    self.repository.deleteMemo(memo: notFixedTasks[indexPath.row])
+                }
             }
+            self.navigationTitle = "\(self.tasks.count)개의 메모"
             self.mainView.tableView.reloadData()
         }
         delete.image = UIImage(systemName: "trash")
+        delete.backgroundColor = .red
         
         return UISwipeActionsConfiguration(actions: [delete])
     }
@@ -79,22 +99,29 @@ extension MemoListViewController: UITableViewDelegate, UITableViewDataSource {
         let notFixedTasks = tasks.filter("isFixed == false")
         
         let fix = UIContextualAction(style: .normal, title: nil) { action, view, completionHandler in
-            if indexPath.section == 0 {
-                self.repository.updateIsFixed(memo: fixedTasks[indexPath.row])
+            if fixedTasks.count == 0 {
+                self.repository.updateIsFixed(memo: notFixedTasks[indexPath.row])
             } else {
-                if fixedTasks.count < 5 {
-                    self.repository.updateIsFixed(memo: notFixedTasks[indexPath.row])
+                if indexPath.section == 0 {
+                    self.repository.updateIsFixed(memo: fixedTasks[indexPath.row])
                 } else {
-                    self.showAlert(title: "알림", message: "고정은 최대 5개까지만 가능합니다!")
+                    if fixedTasks.count < 5 {
+                        self.repository.updateIsFixed(memo: notFixedTasks[indexPath.row])
+                    } else {
+                        self.showAlert(title: "알림", message: "고정은 최대 5개까지만 가능합니다!")
+                    }
                 }
             }
             self.mainView.tableView.reloadData()
         }
-        let image = indexPath.section == 0 ? "pin.slash.fill" : "pin.fill"
-        fix.image = UIImage(systemName: image)
-        
+        if fixedTasks.count == 0 {
+            fix.image = UIImage(systemName: "pin.fill")
+        } else {
+            let image = indexPath.section == 0 ? "pin.slash.fill" : "pin.fill"
+            fix.image = UIImage(systemName: image)
+        }
+        fix.backgroundColor = .systemOrange
         return UISwipeActionsConfiguration(actions: [fix])
-        
     }
 }
